move BlockBuilder to its own file
